Significant changes to searching
The search previously performed a LIKE search across 'target', which was
a normalised field of main, secondary and Maori glosses. The search is
changing to perform a hierarchy of search queries to order results
correctly. Exact matches to major and maori glosses are shown first,
followed by contains matches. After that exact then contains matches are
shown on secondary glosses.

TODO: Normalising of fields (may need to be done in export scripts)