Guard against locationImage() returning nil
Objective-C is more forgiving of values that are nil - it allows more
methods to be called on it. Swift is not as forgiving, so we must guard
against the location image call being nil. It will return nil if there
is no matching location image - at this time, 'Palm' and 'Blades' are
missing locations - I have a query with the client as to whether this is
correct.